chore: add a `lazy` boolean attribute to `RESTObject`
This can be used to tell if a `RESTObject` was created using
`lazy=True`.

Possible future uses for this will be if someone tries to access an
attribute that does not exist they can be notified it is a "lazy"
object, like we do for "list" generated objects.

Another future use could be to give a warning if the attribute that
`_repr_attr` is pointing to is missing and the object is not "lazy".